requirements.txt
kagglehub
pandas
pillow
torch
einops
torchvision
matplotlib
tqdm
nltk